Create endpoints, repository and basic services of the entity Note
In the controller package, the API endpoints were implemented, calling the service layer using dependency injection. The Note repository that interacts with the database wase created in its respective package. Also, it was created a class for the Note Service, which implements an interface that declare the methods the note service needs to implement, these are: saveNote, getNotes, updateNote, and deleteNote. For now, there is not business logic in the service layer implemented yet